move test_extract_value_from_json_response to support
