+++ frame/bridge/eth/offchain/src/tests.rs
@@ -2,36 +2,6 @@ use serde_json;
 // --- darwinia ---
 use crate::{mock::*, *};
 
-/// Extract value from JSON response
-#[test]
-fn test_extract_value_from_json_response() {
-	let result_part =
-		extract_from_json_str(&SUPPOSED_SHADOW_FAKE_RESPONSE[..], b"result" as &[u8]).unwrap();
-	assert_eq!(
-		result_part,
-		br#""eth_header":{eth-content},"proof":[proof-content]"# as &[u8]
-	);
-	let eth_header_part = extract_from_json_str(result_part, b"eth_header" as &[u8]).unwrap();
-	assert_eq!(eth_header_part, br#"eth-content"# as &[u8]);
-	let proof_header_part = extract_from_json_str(result_part, b"proof" as &[u8]).unwrap();
-	assert_eq!(proof_header_part, br#"proof-content"# as &[u8]);
-}
-
-/// Extract value from JSON response with not alphabetical order
-#[test]
-fn test_extract_value_from_non_order_json_response() {
-	let result_part =
-		extract_from_json_str(&SUPPOSED_SHADOW_NON_ORDER_RESPONSE[..], b"result" as &[u8]).unwrap();
-	assert_eq!(
-		result_part,
-		br#""proof":[proof-content],"eth_header":{eth-content}"# as &[u8]
-	);
-	let eth_header_part = extract_from_json_str(result_part, b"eth_header" as &[u8]).unwrap();
-	assert_eq!(eth_header_part, br#"eth-content"# as &[u8]);
-	let proof_header_part = extract_from_json_str(result_part, b"proof" as &[u8]).unwrap();
-	assert_eq!(proof_header_part, br#"proof-content"# as &[u8]);
-}
-
 /// Basice JSON response handle
 #[test]
 fn test_build_eth_header_from_json_response() {
